Update uint64 to float64
float64 bugs

+++ accounting.go
@@ -10,17 +10,17 @@ const (
 type Tx struct {
 	Action TxAction `json:"action"`
 	Signer string   `json:"signer"`
-	Amount uint64   `json:"amount"`
+	Amount float64  `json:"amount"`
 }
 
 type Accounts struct {
 	// Stores the total balance of each account.
-	Accounts map[string]uint64 `json:"accounts"`
+	Accounts map[string]float64 `json:"accounts"`
 }
 
 func newAccounts() *Accounts {
 	return &Accounts{
-		Accounts: make(map[string]uint64),
+		Accounts: make(map[string]float64),
 	}
 }
 
@@ -34,7 +34,7 @@ func (a *Accounts) createAccount(signer string) error {
 	}
 }
 
-func (a *Accounts) balanceOf(signer string) (uint64, error) {
+func (a *Accounts) balanceOf(signer string) (float64, error) {
 	balance, ok := a.Accounts[signer]
 	if ok {
 		return balance, nil
@@ -43,8 +43,8 @@ func (a *Accounts) balanceOf(signer string) (uint64, error) {
 	}
 }
 
-func (a *Accounts) deposit(signer string, amount uint64) *Tx {
-	var newBalance uint64
+func (a *Accounts) deposit(signer string, amount float64) *Tx {
+	var newBalance float64
 
 	balance, err := a.balanceOf(signer)
 	if err != nil {
@@ -62,7 +62,7 @@ func (a *Accounts) deposit(signer string, amount uint64) *Tx {
 	}
 }
 
-func (a *Accounts) withdraw(signer string, amount uint64) (*Tx, error) {
+func (a *Accounts) withdraw(signer string, amount float64) (*Tx, error) {
 	balance, err := a.balanceOf(signer)
 	if err != nil {
 		return nil, &AccountNotFoundError{signer}
@@ -81,7 +81,7 @@ func (a *Accounts) withdraw(signer string, amount uint64) (*Tx, error) {
 	}
 }
 
-func (a *Accounts) send(sender string, recipient string, amount uint64) ([]*Tx, error) {
+func (a *Accounts) send(sender string, recipient string, amount float64) ([]*Tx, error) {
 	wtx, err := a.withdraw(sender, amount)
 	if err != nil {
 		return nil, err
