+++ order_test.go
@@ -10,5 +10,5 @@ func TestNewOrder(t *testing.T) {
 	order := newOrder(Bid, 5)
 
 	assert.Equal(t, Bid, order.Side, "order side should be Bid")
-	assert.Equal(t, uint64(5), order.Size, "order size should be 5")
+	assert.Equal(t, float64(5), order.Size, "order size should be 5")
 }
